backend/requirements.txt
flask==2.2.5
flask-cors==4.0.0
python-dotenv==1.0.0
flasgger==0.9.5
fastapi
uvicorn[standard]
httpx>=0.24,<1
PyJWT==2.8.0
bcrypt==4.0.1
email-validator>=2.0.0